+++ src/main/java/insane96mcp/iguanatweaksreborn/module/misc/Tweaks.java
@@ -309,7 +309,6 @@ public class Tweaks extends Feature {
         Vec3 originalResult = originalOperation.call(instance, pTravelVector, pFriction);
         if (instance.horizontalCollision && !instance.level().isClientSide) {
             double length = horizontalDistance - instance.getDeltaMovement().horizontalDistance();
-            //TODO add a death message like when doomed to fall
             if (length > 0.41f) {
                 instance.hurt(instance.damageSources().source(Tweaks.COLLIDE_WITH_WALL, null), (float) ((length - 0.2f) * collideWithWallsDamage));