+++ changelog.md
@@ -11,6 +11,10 @@
 * Death
   * Killer's names now include the original mob's name
   * Fixed vindication vs killer not working if graves were disabled or keepInventory was enabled
+* Client Misc: Added 'Third Person Death'
+  * When you die, your camera will switch to third person
+  * The death screen has been changed to hide the buttons instead of having them disabled
+  * The buttons are hidden for 2.5 seconds
 * Fixed missing death.attack.collideWithWall.player
 
 ## 3.15.2